do not assume entries made on dos systems have utf-8 filename unless header flag says so
Filenames have been assumed utf-8 and only when decoding as such failed are other encodings taken into account.
Check if a file entry is created on a dos system and if the "Language encoding flag" is cleared, we read filename as
"Code Page 437" text as described in:
    https://pkware.cachefly.net/webdocs/casestudies/APPNOTE.TXT Appendix D

This makes decoding of archives from WinZip with extended latin characters work better.