New writeFileAtPath method controls whether zip64 is used
This new method allows clients to specify whether the zip64 format should be used when writing the specified file. Other methods are unchanged: that is, they always use zip64.

New parameter for SSZipArchive: `useZip64`

The designated initializer for SSZipArchive can now take a `useZip64` parameter to determine whether all the files in the archive should be written with the zip64 format.

The old initializer (-initWithPath:) remains, and creates an archive that uses zip64 (as before).